backend/pyproject.toml
[tool.poetry]
name = "letterboxd-wrapped"
version = "0.1.0"
description = "Letterboxd Wrapped-style analytics"
authors = ["Connor Swartz <connoreswartz@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlmodel = "^0.0.14"
httpx = "^0.25.2"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true